fix(security): secure_remove no-ops on an absent path (Windows); split fs win FFI
Bug: secure_remove documents that removing a missing path is a no-op
success, and the open-for-write step already maps NotFound to Ok. But on
Windows the win_clear_readonly(path)? pre-clear ran first and surfaced
GetFileAttributesW's "file not found" as a hard error, so secure_remove
wrongly returned Err on an absent path. This is why the Tier 2
cargo-mutants baseline failed: the cross-platform test
secure_remove_absent_is_ok panicked on the Windows runner, so
cargo-mutants never reached the mutation phase. Fix: treat a NotFound
attribute-probe result as a no-op; only non-NotFound failures propagate.

The fix nudged fs.rs to 805 LOC, past the 800 file-size gate. Rather than
trim the fix's docs to sneak under (hiding the real issue), extract the
~250 lines of cfg(windows) security FFI helpers (win_set_hidden /
win_clear_readonly / path_to_wide / win_get_file_attributes /
win_set_file_attributes / win_set_owner_only_acl / win_apply_owner_ace)
into a new fs/win.rs module. fs.rs is now 559 LOC and win.rs 264 — both
honestly under the limit with FULL documentation, and the unsafe Win32
calls are isolated in one auditable place. No size exception added.

The regression test (secure_remove_absent_is_ok) already exists and is
cross-platform; it simply is not run on Windows in the PR-fast path.

Validated: file-size policy, host uffs-security suite (49 tests),
Windows-target (xwin) compile.
